requirements.txt
transformers
auto-gptq
py-cord
python-dotenv